# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistem Kasir Enterprise API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: Sales report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Get top selling products
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /reports/revenue-by-category:
    get:
      operationId: getRevenueByCategory
      tags: [reports]
      summary: Revenue breakdown by category
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Revenue by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryRevenue"
  /reports/payment-methods:
    get:
      operationId: getPaymentMethodStats
      tags: [reports]
      summary: Revenue breakdown by payment method
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payment method stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethodStat"
  /reports/hourly-sales:
    get:
      operationId: getHourlySales
      tags: [reports]
      summary: Sales by hour of day
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Hourly sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourlySale"

  # DISCOUNTS
...
components:
  schemas:
...
    SalesDataPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        transactions:
          type: integer
        avgOrderValue:
          type: number
      required:
        - date
        - revenue
        - transactions
        - avgOrderValue
    SalesReport:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        totalRevenue:
          type: number
        totalTransactions:
          type: integer
        avgOrderValue:
          type: number
        data:
          type: array
          items:
            $ref: "#/components/schemas/SalesDataPoint"
      required:
        - from
        - to
        - totalRevenue
        - totalTransactions
        - avgOrderValue
        - data
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        quantitySold:
          type: integer
        revenue:
          type: number
        imageUrl:
          type: ["string", "null"]
      required:
        - productId
        - productName
        - sku
        - quantitySold
        - revenue
    CategoryRevenue:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: string
        revenue:
          type: number
        transactions:
          type: integer
        percentage:
          type: number
      required:
        - categoryName
        - revenue
        - transactions
        - percentage
    PaymentMethodStat:
      type: object
      properties:
        method:
          type: string
        amount:
          type: number
        count:
          type: integer
        percentage:
          type: number
      required:
        - method
        - amount
        - count
        - percentage
    HourlySale:
      type: object
      properties:
        hour:
          type: integer
        revenue:
          type: number
        transactions:
          type: integer
      required:
        - hour
        - revenue
        - transactions

    # DISCOUNT
...